Some text are missing in eject sdcard warning
The eject sdcard warning page shows incomplete texts
in some languages (german) and it is not apparant that
that you can scroll to see the text.

The layout is divided into three parts, the button
and one blank line, and the text. The text and the
blanke space will have equal parts of the screen
since they have the same weight. Because of this the
(complete) text is not shown in som languages.

Remove the blank part and let the text use the whole
screen.

Change-Id: Id8fd88fc6c0aa6105d64c100e3365cd9b00bf851