Dashboard: use own drawable for external apps, if available
In marshmallow, Google added the ability for system apps to populate the Settings dashboard. An infamous example is the Google Play Services "Google" shortcut placed under Language & input.

Following the code you learn that icons are made available by the external apps with intents, and the icons which are loaded from external packages get tinted with framework's colorAccent.

This situation produces 2 problems for themers:
1) They have to create another overlay apk for Google Play Services just to change the icon
2) If they want to use custom icons (.e.g pngs) their image file will get tinted without any exclusion

This commit:
1) checks if the icon name provided by the external app also exists inside the Settins resources.
2) If it exists, it uses that and NOT the external one. Also, it doesn't tint it.
3) Adds the ic_google_settings.xml (the Google shortcut icon):
- I couldn't get the overlay to work without the resource existing in the original package (RRO limitation? it only displayed a rectangle. I think Settings didn't know what to do with the overlay's drawable id.)
- It tints it like the other settings icons (with colorAccent). So the stock theme or a theme which just overlays colorAccent still works.

This way, the original behavior is preserved. Any other (future) shortcuts that Google may add there will still work. If a ROM dev wants to provide the same freedom as with the Google icon,
all they have to do is add the new icon in their Settings. After that, any themer can choose to overlay it or not.